// ProyectoNest-master/src/products/interfaces/product.interface.ts
export interface Product {
        id: string;
        nombre: string;
        marca: string;
        precio: string;
        disponible: boolean;
        modelo: string;
        procesador: string;
        almacenamiento: string;
        ram: string;
        color: string;
        peso: string;
        pantalla: string;
        resolucion: string;
        puertos: string[];
        bateria: string;
        sistema_operativo: string;
        categoria: string;
      }